+++ LehirtiGameEngine/src/org/lehirti/state/IntState.java
@@ -1,5 +1,5 @@
 package org.lehirti.state;
 
 public interface IntState extends State {
-  public Integer defaultValue();
+  public Long defaultValue();
 }
